+++ tests/testthat/test-sps_repweights.R
@@ -12,6 +12,10 @@ test_that("corner cases work as expected", {
     sps_repweights(w, 5, dist = rnorm),
     structure(matrix(1, 10, 5), tau = 1)
   )
+  expect_equal(
+    sps_repweights(w, 5, tau = NULL),
+    structure(matrix(1, 10, 5), tau = 1)
+  )
 
   # asking for 0 repweights gives a matrix with no columns
   expect_equal(
@@ -19,7 +23,7 @@ test_that("corner cases work as expected", {
     structure(matrix(numeric(0), 10, 0), tau = 1)
   )
   expect_equal(
-    sps_repweights(w, 0, dist = rnorm),
+    sps_repweights(w, 0, dist = rnorm, tau = NULL),
     structure(matrix(numeric(0), 10, 0), tau = 1)
   )
 
@@ -29,7 +33,7 @@ test_that("corner cases work as expected", {
     structure(matrix(numeric(0), 0, 5), tau = 1)
   )
   expect_equal(
-    sps_repweights(integer(0), 5, dist = rnorm),
+    sps_repweights(integer(0), 5, dist = rnorm, tau = NULL),
     structure(matrix(numeric(0), 0, 5), tau = 1)
   )
 
@@ -39,7 +43,7 @@ test_that("corner cases work as expected", {
     structure(matrix(numeric(0), 0, 0), tau = 1)
   )
   expect_equal(
-    sps_repweights(integer(0), 0, dist = rnorm),
+    sps_repweights(integer(0), 0, dist = rnorm, tau = NULL),
     structure(matrix(numeric(0), 0, 0), tau = 1)
   )
 })
@@ -86,4 +90,19 @@ test_that("results agree with bootstrapFP:::generalised()", {
   var2 <- bootstrap_fp(y, 1 / w, 100)
 
   expect_equal(var1, var2)
+})
+
+test_that("auto tau works", {
+  set.seed(1234)
+  w <- runif(10) + 1
+  
+  expect_equal(
+    attr(sps_repweights(w, 10, tau = NULL, dist = \(x) rexp(x) - 1), "tau"),
+    1
+  )
+  
+  set.seed(12345)
+  tau <- attr(sps_repweights(w, 20, tau = NULL), "tau")
+  set.seed(12345)
+  expect_warning(sps_repweights(w, 20, tau = tau - 0.001))
 })