+++ NEWS.md
@@ -3,6 +3,10 @@
 - `inclusion_prob()` is now faster with a single stratum, improving the performance
 of `sps()` and `expected_coverage()`.
 
+- Setting `tau = NULL` in `sps_repweights()` automatically picks the smallest
+`tau` to keep the replicate weights from being negative. Inspired by
+`svrep::make_gen_boot_factors()`.
+
 - Now requires R >= 4.2, which has been the case for a while (closing #2).
 
 ## Changes in version 0.5.4